app-restapi/src/requirements.txt
asgiref
Django
psycopg2-binary
pytz
sqlparse
django-crispy-forms
celery
redis
djangorestframework
Markdown
django-filter
pyjwt